+++ src/rotta_rs_module/arrayy/function/abs.rs
@@ -5,7 +5,7 @@ pub fn abs_arr(x: &Arrayy) -> Arrayy {
     let vector = x.value
         .iter()
         .map(|v| v.abs())
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(x.shape.clone(), vector)
 }
