+++ src/rotta_rs_module/arrayy/function/exp.rs
@@ -4,7 +4,7 @@ pub fn exp_arr(arr: &Arrayy) -> Arrayy {
     let vector = arr.value
         .iter()
         .map(|v| { v.exp() })
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(arr.shape.clone(), vector)
 }
