+++ src/rotta_rs_module/arrayy/function/power.rs
@@ -4,16 +4,16 @@ pub fn powi_arr(arr: &Arrayy, n: i32) -> Arrayy {
     let vector = arr.value
         .iter()
         .map(|v| v.powi(n))
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(arr.shape.clone(), vector)
 }
 
-pub fn powf_arr(arr: &Arrayy, n: f64) -> Arrayy {
+pub fn powf_arr(arr: &Arrayy, n: f32) -> Arrayy {
     let vector = arr.value
         .iter()
         .map(|v| v.powf(n))
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(arr.shape.clone(), vector)
 }
