+++ src/rotta_rs_module/arrayy/function/sin_cos_tan.rs
@@ -4,7 +4,7 @@ pub fn sin_arr(arr: &Arrayy) -> Arrayy {
     let vector = arr.value
         .iter()
         .map(|x| { x.sin() })
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(arr.shape.clone(), vector)
 }
@@ -13,7 +13,7 @@ pub fn cos_arr(arr: &Arrayy) -> Arrayy {
     let vector = arr.value
         .iter()
         .map(|x| { x.cos() })
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(arr.shape.clone(), vector)
 }
@@ -22,7 +22,7 @@ pub fn tan_arr(arr: &Arrayy) -> Arrayy {
     let vector = arr.value
         .iter()
         .map(|x| { x.tan() })
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(arr.shape.clone(), vector)
 }
