+++ src/rotta_rs_module/arrayy/function/sum.rs
@@ -1,5 +1,5 @@
 use crate::rotta_rs_module::arrayy::*;
 
-pub fn sum_arr(arr: &Arrayy) -> f64 {
-    arr.value.iter().sum::<f64>()
+pub fn sum_arr(arr: &Arrayy) -> f32 {
+    arr.value.iter().sum::<f32>()
 }
