+++ src/rotta_rs_module/arrayy/operation/math/slice_operation/sub.rs
@@ -1,9 +1,9 @@
 use crate::rotta_rs_module::arrayy::*;
 
 pub fn sub_arr_slice(
-    arr_a: (&[f64], &[usize]),
-    arr_b: (&[f64], &[usize])
-) -> (Vec<f64>, Vec<usize>) {
+    arr_a: (&[f32], &[usize]),
+    arr_b: (&[f32], &[usize])
+) -> (Vec<f32>, Vec<usize>) {
     let (arr_a, shape_a) = arr_a;
     let (arr_b, shape_b) = arr_b;
 
@@ -17,7 +17,7 @@ pub fn sub_arr_slice(
                 // add
                 a - arr_b[i]
             })
-            .collect::<Vec<f64>>();
+            .collect::<Vec<f32>>();
         (vector, shape_a.to_vec())
     } else if shape_a.multiple_sum() == 1 || shape_b.multiple_sum() == 1 {
         // skalar
@@ -28,7 +28,7 @@ pub fn sub_arr_slice(
             let vector = arr_b
                 .iter()
                 .map(|v| skalar - *v)
-                .collect::<Vec<f64>>();
+                .collect::<Vec<f32>>();
 
             (vector, broadcasting_shape)
         } else {
@@ -37,7 +37,7 @@ pub fn sub_arr_slice(
             let vector = arr_a
                 .iter()
                 .map(|v| *v - skalar)
-                .collect::<Vec<f64>>();
+                .collect::<Vec<f32>>();
 
             (vector, broadcasting_shape)
         }
@@ -55,7 +55,7 @@ pub fn sub_arr_slice(
                 // add
                 a - arr_b[i]
             })
-            .collect::<Vec<f64>>();
+            .collect::<Vec<f32>>();
         (vector, broadcast_shape)
     }
 }
