+++ src/rotta_rs_module/module/initialization/glorot.rs
@@ -4,8 +4,8 @@ use rand_distr::Normal;
 use crate::rotta_rs_module::Module;
 
 impl Module {
-    pub fn glorot_initialization(&mut self, input: usize, output: usize) -> f64 {
-        let glorot_formula = 2.0 / ((input + output) as f64);
+    pub fn glorot_initialization(&mut self, input: usize, output: usize) -> f32 {
+        let glorot_formula = 2.0 / ((input + output) as f32);
         let glorot_std = glorot_formula.sqrt();
 
         let normal = Normal::new(0.0, glorot_std).unwrap();
